docs/test: round-3 polish — wording accuracy, fixture readability, co-location (review round 3)
Round-3 NITs, all four lenses: adapter.go's opening parenthetical no
longer over-claims non-command refusal for every implementor (walked
back 22 lines later — now '(unmodeled fields; on most adapters also
non-command handlers)'); architecture.md's dangling wrapped line
reflowed; codex/hook.go's codexHookEvents doc no longer says hooks.json
while render targets config.toml; the codex import test comment carries
its load-bearing difference (MergeTOML whole-file rewrite makes
byte-equality a stronger detector) instead of review provenance; codex
guard fixtures are raw multiline TOML literals instead of escaped
one-liners; the codex captured-verbatim row also asserts Matcher; the
relative fetcher's copy error names the user-typed path alongside the
resolved one; claude's two hook tables now co-locate in
hook_guard_test.go matching the gemini/cursor/codex layout.
